STYLE: Avoid old, C-style cast
Avoiding old-style casts on advice of:
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#es48-avoid-casts
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#es49-if-you-must-use-a-cast-use-a-named-cast
https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-cstyle-cast.html

+++ Examples/Statistics/NormalVariateGenerator.cxx
@@ -51,7 +51,7 @@ main()
   // Software Guide : BeginCodeSnippet
   using GeneratorType = itk::Statistics::NormalVariateGenerator;
   auto generator = GeneratorType::New();
-  generator->Initialize((int)2003);
+  generator->Initialize(2003);
 
   for (unsigned int i = 0; i < 50; ++i)
   {
